src/third_party/WebKit/LayoutTests/platform/mac/external/wpt/html/rendering/non-replaced-elements/the-fieldset-element-0/legend-grid-flex-multicol-expected.txt

This is a testharness.js-based test.
FAIL flex assert_equals: offsetHeight expected 18 but got 36
FAIL inline-flex assert_equals: offsetHeight expected 18 but got 36
FAIL grid assert_equals: offsetHeight expected 18 but got 36
FAIL inline-grid assert_equals: offsetHeight expected 18 but got 36
PASS multicol
Harness: the test ran to completion.
